API/backend/requirements.txt
flask
flask_cors
python-dotenv
python-keycloak
flask-jwt-extended